backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation and Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP Client
httpx==0.26.0

# Gemini API
google-genai>=1.0.0

# Caching
cachetools==5.3.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-snapshot==0.9.0

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
